Update ExtensionStats module to work with new Extensions Manager
Download statistics from the old Midas server is stored in ExtensionsDownloadStats-20210521.csv file and added to the numbers retrieved from the new server.

ExtensionsDownloadStats-20210521.csv is a snapshot that was created a few month before switching to the new extensions server, therefore download counts of post-4.11.20210226 may miss a few months of downloads.